Fix named export conflict in @glimmer/opcode-compiler
Recently a function called `debug` was added to the opcode compiler's `lib/compiler.ts` file.

Unfortunately, the package root was using a wildcard to re-export all of the compiler file's exports. This conflicted with the already existing debug function that was being exported.

Because there existed two exports of the same name, the CommonJS build was failing because the transpiled output uses Object.defineProperty to define non-configurable properties, so duplicates cause a runtime error.

This commit resolves the problem by renaming one of the debug functions to `debugCompiler`. It also removes any wildcard exports from the package entry point to prevent similar issues from happening in the future.